+++ src/Core/Configuration.lua
@@ -91,8 +91,28 @@ local Configuration = {}
     @local
     ]]
     function Configuration:handle(...)
-        -- @TODO: Document param possibilities <2024.04.22>
-        -- @TODO: Implement this method <2024.04.22>
+        if self.data == nil then
+            self.__.output:out('There was an attempt to get or set configuration values with no addon respective data set. Please, pass the data variable name when initializing the Stormwind Library to use this feature.')
+            return nil
+        end
+
+        local arg1, arg2, arg3 = ...
+
+        if type(arg1) == 'string' then
+            if self.__.bool:isTrue(arg3) then
+                return self:getOrInitialize(arg1, arg2)
+            else
+                return self:get(arg1, arg2)
+            end
+        end
+
+        if type(arg1) == 'table' then
+            self.__.arr:each(arg1, function(value, key)
+                self:set(key, value)
+            end)
+        end
+
+        return nil
     end
 
     --[[--